Move todo files out of directories
Moving todo files out of directories in order to match the style of the user files.
