+++ client/src/app/app.module.ts
@@ -31,9 +31,9 @@ import { HttpClientModule } from '@angular/common/http';
 import { LayoutModule } from '@angular/cdk/layout';
 import { UserCardComponent } from './users/user-card.component';
 import { UserProfileComponent } from './users/user-profile.component';
-import { TodosCardComponent } from './todos/todos-card/todos-card.component';
-import { TodosListComponent } from './todos/todos-list/todos-list.component';
-import { TodosProfileComponent } from './todos/todos-profile/todos-profile.component';
+import { TodosCardComponent } from './todos/todos-card.component';
+import { TodosListComponent } from './todos/todos-list.component';
+import { TodosProfileComponent } from './todos/todos-profile.component';
 
 const MATERIAL_MODULES: any[] = [
   MatListModule,
